+++ src/pages/MyPage/AddIdol/compoents/IdolCard.jsx
