+++ src/pages/MyPage/AddIdol/compoents/SlideButton.jsx
