+++ src/pages/MyPage/AddIdol/compoents/SlideButton.module.css
