+++ src/pages/MyPage/InterestIdol/components/InterestIdol.module.css
