+++ src/pages/listPage/Chart/components/WarningModal.module.css
@@ -1,0 +1,72 @@
+.modalOverlay {
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100vw;
+  height: 100vh;
+  background: rgba(0, 0, 0, 0.5);
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  z-index: 999;
+}
+
+.modalContainer {
+  background: #181d26;
+  border-radius: 12px;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  padding: 24px;
+  position: relative;
+}
+
+.icon {
+  width: 180px;
+  height: 180px;
+}
+
+.closeButton {
+  position: absolute;
+  top: 24px;
+  right: 20px;
+  background: none;
+  border: none;
+  cursor: pointer;
+  z-index: 1001;
+}
+
+.closeIcon {
+  width: 24px;
+  height: 24px;
+}
+
+.messageText {
+  color: #ffffff;
+  font-size: 16px;
+  font-weight: 500;
+  text-align: center;
+  margin-bottom: 16px;
+}
+
+.highlight {
+  color: #f96d69;
+}
+
+.confirmButton {
+  width: 295px;
+  height: 42px;
+  background: linear-gradient(90deg, #f86f65, #fe5493);
+  color: #ffffff;
+  border: none;
+  border-radius: 8px;
+  cursor: pointer;
+  font-size: 16px;
+  margin-top: 20px;
+  margin-bottom: 10px;
+}
+
+.confirmButton:hover {
+  background: linear-gradient(90deg, #ff5252, #ff3b3b);
+}
